+++ cmd/tailscaled/depaware.txt
@@ -279,9 +279,9 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
    L    tailscale.com/ipn/store/awsstore                             from tailscale.com/ipn/store
    L    tailscale.com/ipn/store/kubestore                            from tailscale.com/ipn/store
         tailscale.com/ipn/store/mem                                  from tailscale.com/ipn/ipnlocal+
-   L    tailscale.com/kube/api                                       from tailscale.com/ipn/store/kubestore+
-   L    tailscale.com/kube/client                                    from tailscale.com/ipn/store/kubestore
-        tailscale.com/kube/types                                     from tailscale.com/envknob
+   L    tailscale.com/kube/kubeapi                                   from tailscale.com/ipn/store/kubestore+
+   L    tailscale.com/kube/kubeclient                                from tailscale.com/ipn/store/kubestore
+        tailscale.com/kube/kubetypes                                 from tailscale.com/envknob
         tailscale.com/licenses                                       from tailscale.com/client/web
         tailscale.com/log/filelogger                                 from tailscale.com/logpolicy
         tailscale.com/log/sockstatlog                                from tailscale.com/ipn/ipnlocal
