+++ ipn/store/kubestore/store_kube.go
@@ -13,8 +13,8 @@ import (
 	"time"
 
 	"tailscale.com/ipn"
-	kubeapi "tailscale.com/kube/api"
-	kubeclient "tailscale.com/kube/client"
+	"tailscale.com/kube/kubeapi"
+	"tailscale.com/kube/kubeclient"
 	"tailscale.com/types/logger"
 )
