+++ kube/kubeclient/client.go
@@ -1,13 +1,13 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
-// Package client provides a client to interact with Kubernetes.
+// Package kubeclient provides a client to interact with Kubernetes.
 // This package is Tailscale-internal and not meant for external consumption.
 // Further, the API should not be considered stable.
 // Client is split into a separate package for consumption of
 // non-Kubernetes shared libraries and binaries. Be mindful of not increasing
 // dependency size for those consumers when adding anything new here.
-package client
+package kubeclient
 
 import (
 	"bytes"
@@ -26,7 +26,7 @@ import (
 	"sync"
 	"time"
 
-	kubeapi "tailscale.com/kube/api"
+	"tailscale.com/kube/kubeapi"
 	"tailscale.com/util/multierr"
 )
