+++ kube/kubetypes/grants.go
@@ -1,12 +1,12 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
-// Package types contains types and constants related to the Tailscale
+// Package kubetypes contains types and constants related to the Tailscale
 // Kubernetes Operator.
 // These are split into a separate package for consumption of
 // non-Kubernetes shared libraries and binaries. Be mindful of not increasing
 // dependency size for those consumers when adding anything new here.
-package types
+package kubetypes
 
 import "net/netip"
